\name{sample.int}
\alias{sample.int}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Random Samples and Permutations
%%
}
\description{
%%  sample.int takes a sample of the specified size from the elements of x with replacement.
}
\usage{sample.int(n, size = n, replace = TRUE, prob = NULL)
}


\details{see sample {base}
%%
}
\value{see sample {base}
%%  see sample {base}
}
\references{see sample {base}
%% see sample {base}
}
\author{rcp
%%  rcp
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }% use one of  RShowDoc("KEYWORDS")
